backend/requirements.txt
anthropic>=0.85.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.4.2
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
youtube-transcript-api==0.6.2
pydub==0.25.1
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.29.0
